userUserOrRedirect
