+++ src/app/components/queue/QueueItem.tsx
@@ -1,11 +1,12 @@
 import { IdentifiableQuestion, IdentifiableUsers } from "@interfaces/type";
 import { Grid, Typography } from "@mui/material";
 import ArrowRightOutlinedIcon from "@mui/icons-material/ArrowRightOutlined";
-import { getUserSessionOrRedirect, trimUserName } from "@utils/index";
+import { trimUserName } from "@utils/index";
 import React from "react";
 import { getUsers } from "@services/client/user";
 import { useRouter } from "next/navigation";
 import { useOfficeHour } from "@hooks/oh/useOfficeHour";
+import { useUserOrRedirect } from "@hooks/useUserOrRedirect";
 
 interface QueueItemProps {
   order: number;
@@ -19,7 +20,7 @@ const QueueItem = (props: QueueItemProps) => {
 
   const router = useRouter();
   const { course } = useOfficeHour();
-  const user = getUserSessionOrRedirect();
+  const user = useUserOrRedirect();
 
   if (!user) {
     return null;
