+++ src/app/hooks/useUserOrRedirect.tsx
@@ -1,0 +1,15 @@
+import { useUserSession } from "@context/UserSessionContext";
+import { useRouter } from "next/navigation";
+import React from "react";
+
+export const useUserOrRedirect = () => {
+  const { user, session } = useUserSession();
+  const router = useRouter();
+  React.useEffect(() => {
+    if (!user && !session.isLoading) {
+      router.push("/login");
+    }
+  }, [user, router, session]);
+
+  return user ?? null;
+};
