+++ src/app/login/page.tsx
@@ -43,7 +43,7 @@ export default function Page() {
         </div>
         <div className="mt-1">Answers in minutes, Not hours</div>
         <Button
-          className={`py-2.5 px-6 bg-[#38608F] rounded-full mt-6 normal-case `}
+          className="py-2.5 px-6 bg-[#38608F] rounded-full mt-6 normal-case"
           onClick={() => onSignIn()}
         >
           <Typography variant="subtitle2" color="#FFFFFF" fontWeight={600}>
