+++ src/app/private/course/[courseId]/board/history/page.tsx
@@ -2,9 +2,10 @@
 import Board from "@components/board";
 import Header from "@components/Header";
 import { useOfficeHour } from "@hooks/oh/useOfficeHour";
+import { useUserOrRedirect } from "@hooks/useUserOrRedirect";
 import { ArrowBack } from "@mui/icons-material";
 import { Box, Typography } from "@mui/material";
-import { getExpiredQuestions, getUserSessionOrRedirect } from "@utils/index";
+import { getExpiredQuestions } from "@utils/index";
 import Link from "next/link";
 import React from "react";
 
@@ -17,7 +18,7 @@ interface PageProps {
 const Page = (props: PageProps) => {
   const { courseId } = props.params;
   const { course, questions } = useOfficeHour();
-  const user = getUserSessionOrRedirect();
+  const user = useUserOrRedirect();
   if (!user) {
     return null;
   }
