+++ src/app/private/course/[courseId]/board/page.tsx
@@ -3,11 +3,9 @@
 import Board from "@components/board";
 import Header from "@components/Header";
 import { useOfficeHour } from "@hooks/oh/useOfficeHour";
+import { useUserOrRedirect } from "@hooks/useUserOrRedirect";
 import { Box, Button, Typography } from "@mui/material";
-import {
-  getActivePublicQuestion,
-  getUserSessionOrRedirect,
-} from "@utils/index";
+import { getActivePublicQuestion } from "@utils/index";
 import Link from "next/link";
 
 import React from "react";
@@ -22,7 +20,7 @@ const Page = (props: PageProps) => {
   const { courseId } = props.params;
 
   const { course, questions } = useOfficeHour();
-  const user = getUserSessionOrRedirect();
+  const user = useUserOrRedirect();
   if (!user) {
     return null;
   }
